initcpio: Move firmware unpacking to the initramfs
This solves the flakines due to udev getting uevents for devices before
the firmware is updated, and gets rid of the dodgy systemd units
involved. Now the initramfs is in charge of unpacking firmware twice:
once for itself, once to update the rootfs firmware (if required).

This is all fast enough that there isn't any significant impact on boot
time.

Signed-off-by: Hector Martin <marcan@marcan.st>