asahi-diagnose: Use hostnamectl to fetch hostname
A default installation of Asahi Linux doesn't appear to come included with the `hostname` binary. This causes the hostname lookup to fail when asahi-diagnose is used.

This technically ties the script to systemd, so a better solution if this becomes a concern (i.e. if different init systems are offered in the future) might be to check for the presence of the `hostname` binary, and if it's not present, go with a sensible alternative value instead of erroring out.

Signed-off-by: Toby Jones <github@monotron.me>