dracut: conf: Fix i2c module name
Fixes USB support in the initramfs by reverting the unintentional
omission of apple-dart.ko, apple-rtkit-helper.ko,
spmi-apple-controller.ko and dwc3-of-simple.ko from the initramfs. Due
to the missing dart module lot's of other HW was unusable in the
initramfs as well.
Does not fix the late probing of the rtc-macsmc modfule.

The upstream module name for the i2c blocks is i2c-pasemi-platform.
Although dracut includes the module already it on its own it still
causes problems. Dracut seems to ignore all other modules when it can't
find one of the entries in the 'add_drivers' variable.

Signed-off-by: Janne Grunau <j@jannau.net>