dracut: conf: Use "drivers" instead of "add_drivers"
The only difference between those to variables is that "add_drivers"
checks for the existance of the module and fails to process further
modules. This is however just logged and not a fatal error for the
initramfs creation. So the verification is worse than useless.
No verification is done for 'drivers' so we can use that even to handle
the i2c module rename in Linux 6.8.6.

Fixes: f444f0a ("dracut: conf: Fix i2c module name")
Signed-off-by: Janne Grunau <j@jannau.net>