update-m1n1: Drop arch specific default DTBS glob pattern
The glob pattern for the Apple silicon devicetrees is OS dependent. Do
not pretend that's possible to provide a default value and instead error
out if it is unset.
Motivated by the addition of devicetrees for non macOS devices which
should be excluded from the 2nd stage m1n1 binary.

Signed-off-by: Janne Grunau <j@jannau.net>