Install hwdb file to override auto suspend on GL9755 SD card reader
systemd v255 shipped a hwdb update enabling auto suspend (runtime PM)
for the SD card reader Apple uses in 14- and 16-inch Macbook Pros and
Mac Studios
(https://github.com/systemd/systemd/commit/2f72e94984fc2994ab2705e384f92e15cc026264).
This PCI ID is imported from ChromiumOS. The driver in upstream Linux
does not support runtime PM so this change is questionable to useless.
Unfortunately it creates a race between enabling auto suspend and probing
the sdhci_pci driver. If the default idle timeout (2000 ms) triggers
before the driver is bound the default PCI device runtime suspend
callback will be called. On Apple silicon devices the GL9755 card
reader is inaccessible after it is resumed for the driver's probe
function. Accessing the device's BAR results in a SError with
ESR = 0x92000018.
It's not entirely clear whether which is an issue with the device or
with Apple silicon's PCIe implementation or Linux driver.

Signed-off-by: Janne Grunau <j@jannau.net>