dracut: Add apple_spmi_nvmem module
The SPMI nvmem support was merged as apple_spmi_nvmem for v6.16 and
Asahi's v6.15 based downstream tree pick those patches up. Add the
module the dracut config to ensure it ends up in the initramfs.
Keep the old modules until v6.16++ is released to ensure compatibility
with older kernels for a while.

Signed-off-by: Janne Grunau <j@jannau.net>